started working on router and various tools
