+++ src/components/calculator/stats-calculator/other-line/OtherLine.tsx
