+++ src/components/calculator/stats-calculator/resist-line/ResLine.tsx
