+++ src/styles/_alert-box.scss
@@ -66,8 +66,8 @@
 
 .spinner {
   margin: 0 auto;
-  border: 16px solid #f3f3f3; /* Light grey */
-  border-top: 16px solid #3498db; /* Blue */
+  border: 16px solid #9ea33f;
+  border-top: 16px solid #b9bd6f;
   border-radius: 50%;
   width: 40px;
   height: 40px;
